changes based on comments in PR https://github.com/katholiek-onderwijs-vlaanderen/sri4node-attachments/pull/19 + some more test cases
