+++ .eslintrc.js
@@ -1,19 +1,18 @@
 module.exports = {
-  env: {
-    commonjs: true,
-    es2021: true,
-    node: true,
-    mocha: true,
-  },
-  extends: ["eslint:recommended", "prettier"],
-  overrides: [],
-  parserOptions: {
-    ecmaVersion: "latest",
-  },
-  rules: {
-    "no-unused-vars": [
-      "error",
-      { args: "all", varsIgnorePattern: "^_", argsIgnorePattern: "^_" },
+    "env": {
+        "browser": true,
+        "commonjs": true,
+        "es2021": true,
+        "mocha": true,
+    },
+    "extends": "eslint:recommended",
+    "overrides": [
     ],
-  },
+    "parserOptions": {
+        "ecmaVersion": "latest"
+    },
+    "rules": {
+      "no-unused-vars": ["error", { "args": "all", "varsIgnorePattern": "^_", "argsIgnorePattern": "^_" }],
+      "semi": [2, "always"],
+    }
 };
