+++ test/tests.js
@@ -5,7 +5,7 @@ const sleep = require('await-sleep');
 
 const sri4nodeConfigFactory = require("./context/config");
 
-const testPartyAttachmentsAfterUpdateMod =  require("./testPartyAttachmentsAfterUpdate");
+const testPartyAttachmentsCheckStoreAttachmentMod =  require("./testPartyAttachmentsCheckStoreAttachment");
 
 const port = 5000;
 const base = `http://localhost:${port}`;
@@ -15,196 +15,144 @@ const httpClient = httpClientMod.httpClientFactory(base);
 
 const { info, error } = require("../js/common");
 
-describe("sri4node-attachments : ", () => {
-  /** @type {import("sri4node").TSriServerInstance} */
-  let sri4nodeServerInstance;
-  let server;
+let serverStarted = false;
 
-  before(async () => {
-    try {
-      const app = expressFactory();
-      const sriConfig = await sri4nodeConfigFactory();
-      sriConfig.description = 'config of sri4node-attachments';
-      sri4nodeServerInstance = await sri4node.configure(app, sriConfig);
-      app.set("port", port);
-      server = app.listen(port, () => {
-        info(`Node app is running at localhost:${port}`);
-      });
-    } catch (err) {
-      error("Unable to start server.");
-      error(err);
-      error(err.stack);
-    }
-  });
-
-  after(async () => {
-    // enable this to keep the server running for inspection
-    // await new Promise(() => {});
-    try {
-      await server.close();
-    } catch (err) {
-      console.log("Closing express server failed");
-    }
-    try {
-      await sri4nodeServerInstance.pgp.end();
-    } catch (err) {
-      console.log("Closing sri4nodeServerInstance failed");
-    }
-  });
-
-  require("./unitTests");
-  require("./testPartyAttachments")(httpClient, "/partiesS3");
-
-  // local storage is currenlty not supported anymore
-  // require("./testPartyAttachments")(base, "/partiesFolder");
-});
-
-
-describe("sri4node-attachments(2) : ", () => {
-  /** @type {import("sri4node").TSriServerInstance} */
-  let sri4nodeServerInstance;
-  let server;
-
-  before(async () => {
-    try {
+/**
+ * 
+ * @param {string} id 
+ * @param {boolean} handleMultipleUploadsTogether 
+ * @param {boolean} uploadInSequence 
+ * @param {*} customStoreAttachment 
+ * @returns 
+ */
+const initServer = async (id, handleMultipleUploadsTogether, uploadInSequence, customStoreAttachment) => {
+  try {
+    const app = expressFactory();
+    if (serverStarted) {
       // It seems that Express has no method to deinitialize or to clear routes.
       // Workaround: let the 'app' variable go out of scope and wait 5 seconds, this
       // seems to deinitialize Express. If Express is just reinitiated with its new
       // configuration without some waiting before usage, Express keeps using the old
       // routes (probably somehow cached).
-      const app = expressFactory();
       await sleep(5000);
-
-      const handleMultipleUploadsTogether = false;
-      const uploadInSequence = false;
-      const sriConfig = await sri4nodeConfigFactory(handleMultipleUploadsTogether, uploadInSequence, testPartyAttachmentsAfterUpdateMod.checkStoreAttachmentFactory(handleMultipleUploadsTogether, uploadInSequence));
-      sriConfig.description = 'config of sri4node-attachments(2)';
-      sri4nodeServerInstance = await sri4node.configure(app, sriConfig);
-      app.set("port", port);
-      server = app.listen(port, () => {
-        info(`Node app is running at localhost:${port}`);
-      });
-    } catch (err) {
-      error("Unable to start server.");
-      error(err);
-      error(err.stack);
     }
-  });
 
-  after(async () => {
-    // enable this to keep the server running for inspection
-    // await new Promise(() => {});
-    try {
-      await server.close();
-    } catch (err) {
-      console.log("Closing express server failed");
-    }
-    try {
-      await sri4nodeServerInstance.pgp.end();
-    } catch (err) {
-      console.log("Closing sri4nodeServerInstance failed");
-    }
-  });
+    const sriConfig = await sri4nodeConfigFactory(handleMultipleUploadsTogether, uploadInSequence, customStoreAttachment);
+    sriConfig.description = `config of sri4node-attachments(${id})`;
+    const sri4nodeServerInstance = await sri4node.configure(app, sriConfig);
+    app.set("port", port);
+    const server = app.listen(port, () => {
+      info(`Node app is running at localhost:${port}`);
+    });
+    serverStarted = true;
+    return { sri4nodeServerInstance, server };
+  } catch (err) {
+    error("Unable to start server.");
+    error(err);
+    error(err.stack);
+    process.exit(1);
+  }
+};
 
-  testPartyAttachmentsAfterUpdateMod.factory(httpClient, "/partiesS3");
+const closeServer = async (server, sri4nodeServerInstance) => {
+  try {
+    await server.close();
+  } catch (err) {
+    console.log("Closing express server failed");
+  }
+  try {
+    await sri4nodeServerInstance.pgp.end();
+  } catch (err) {
+    console.log("Closing sri4nodeServerInstance failed");
+  }
+};
+
+const runTests = async (httpClient, checkStoreAttachmentsReceivedList) => {
+  require("./testPartyAttachments")(httpClient, "/partiesS3");
+
+  testPartyAttachmentsCheckStoreAttachmentMod.factory(httpClient, "/partiesS3", checkStoreAttachmentsReceivedList);
+
+  // local storage is currenlty not supported anymore
+  // require("./testPartyAttachments")(base, "/partiesFolder");
+};
+
+
+describe("Unit tests : ", () => {
+  require("./unitTests");
 });
 
-describe("sri4node-attachments(3) : ", () => {
+// To be able to test the attachments plugin with different configuration parameters we need to start
+// different server instances:
+//                                    handleMultipleUploadsTogether  uploadInSequence
+//   sri4node-attachments(1) :                  false                       n/a
+//   sri4node-attachments(2) :                  true                        true
+//   sri4node-attachments(3) :                  true                        false
+
+
+describe("sri4node-attachments(1) : ", () => {
   /** @type {import("sri4node").TSriServerInstance} */
   let sri4nodeServerInstance;
   let server;
   const checkStoreAttachmentsReceivedList = [];
 
   before(async () => {
-    try {
-      // It seems that Express has no method to deinitialize or to clear routes.
-      // Workaround: let the 'app' variable go out of scope and wait 5 seconds, this
-      // seems to deinitialize Express. If Express is just reinitiated with its new
-      // configuration without some waiting before usage, Express keeps using the old
-      // routes (probably somehow cached).
-      const app = expressFactory();
-      await sleep(5000);
-
-      const handleMultipleUploadsTogether = false;
-      const uploadInSequence = true;
-      const sriConfig = await sri4nodeConfigFactory(handleMultipleUploadsTogether, uploadInSequence, testPartyAttachmentsAfterUpdateMod.checkStoreAttachmentFactory(handleMultipleUploadsTogether, uploadInSequence, checkStoreAttachmentsReceivedList));
-      sriConfig.description = 'config of sri4node-attachments(3)';
-      sri4nodeServerInstance = await sri4node.configure(app, sriConfig);
-      app.set("port", port);
-      server = app.listen(port, () => {
-        info(`Node app is running at localhost:${port}`);
-      });
-    } catch (err) {
-      error("Unable to start server.");
-      error(err);
-      error(err.stack);
-    }
+    const handleMultipleUploadsTogether = false;
+    const uploadInSequence = false; // uploadInSequence value does not matter in case of handleMultipleUploadsTogether=false
+                                    // (not relevant and thus not used in that code path)
+    const customStoreAttachment = testPartyAttachmentsCheckStoreAttachmentMod.checkStoreAttachmentFactory(handleMultipleUploadsTogether, uploadInSequence, checkStoreAttachmentsReceivedList);
+    ({ sri4nodeServerInstance, server } = await initServer('1', handleMultipleUploadsTogether, uploadInSequence, customStoreAttachment));
   });
 
   after(async () => {
     // enable this to keep the server running for inspection
     // await new Promise(() => {});
-    try {
-      await server.close();
-    } catch (err) {
-      console.log("Closing express server failed");
-    }
-    try {
-      await sri4nodeServerInstance.pgp.end();
-    } catch (err) {
-      console.log("Closing sri4nodeServerInstance failed");
-    }
+    await closeServer(server, sri4nodeServerInstance);
   });
 
-  testPartyAttachmentsAfterUpdateMod.factory(httpClient, "/partiesS3", checkStoreAttachmentsReceivedList);
+  runTests(httpClient, checkStoreAttachmentsReceivedList);
 });
 
 
-describe("sri4node-attachments(4) : ", () => {
+// Configuration with multiple uploads together
+describe("sri4node-attachments(2) : ", () => {
   /** @type {import("sri4node").TSriServerInstance} */
   let sri4nodeServerInstance;
   let server;
 
   before(async () => {
-    try {
-      // It seems that Express has no method to deinitialize or to clear routes.
-      // Workaround: let the 'app' variable go out of scope and wait 5 seconds, this
-      // seems to deinitialize Express. If Express is just reinitiated with its new
-      // configuration without some waiting before usage, Express keeps using the old
-      // routes (probably somehow cached).
-      const app = expressFactory();
-      await sleep(5000);
-
-      const handleMultipleUploadsTogether = true;
-      const uploadInSequence = true;
-      const sriConfig = await sri4nodeConfigFactory(handleMultipleUploadsTogether, uploadInSequence, testPartyAttachmentsAfterUpdateMod.checkStoreAttachmentFactory(handleMultipleUploadsTogether, uploadInSequence));
-      sriConfig.description = 'config of sri4node-attachments(4)';
-      sri4nodeServerInstance = await sri4node.configure(app, sriConfig);
-      app.set("port", port);
-      server = app.listen(port, () => {
-        info(`Node app is running at localhost:${port}`);
-      });
-    } catch (err) {
-      error("Unable to start server.");
-      error(err);
-      error(err.stack);
-    }
+    const handleMultipleUploadsTogether = true;
+    const uploadInSequence = true;
+    const customStoreAttachment = testPartyAttachmentsCheckStoreAttachmentMod.checkStoreAttachmentFactory(handleMultipleUploadsTogether, uploadInSequence);
+    ({ sri4nodeServerInstance, server } = await initServer('2', handleMultipleUploadsTogether, uploadInSequence, customStoreAttachment));
   });
 
   after(async () => {
     // enable this to keep the server running for inspection
     // await new Promise(() => {});
-    try {
-      await server.close();
-    } catch (err) {
-      console.log("Closing express server failed");
-    }
-    try {
-      await sri4nodeServerInstance.pgp.end();
-    } catch (err) {
-      console.log("Closing sri4nodeServerInstance failed");
-    }
+    await closeServer(server, sri4nodeServerInstance);
   });
 
-  testPartyAttachmentsAfterUpdateMod.factory(httpClient, "/partiesS3");
+  runTests(httpClient);
+});
+
+// Configuration with multiple uploads together
+describe("sri4node-attachments(3) : ", () => {
+  /** @type {import("sri4node").TSriServerInstance} */
+  let sri4nodeServerInstance;
+  let server;
+
+  before(async () => {
+    const handleMultipleUploadsTogether = true;
+    const uploadInSequence = false;
+    const customStoreAttachment = testPartyAttachmentsCheckStoreAttachmentMod.checkStoreAttachmentFactory(handleMultipleUploadsTogether, uploadInSequence);
+    ({ sri4nodeServerInstance, server } = await initServer('3', handleMultipleUploadsTogether, uploadInSequence, customStoreAttachment));
+  });
+
+  after(async () => {
+    // enable this to keep the server running for inspection
+    // await new Promise(() => {});
+    await closeServer(server, sri4nodeServerInstance);
+  });
+
+  runTests(httpClient);
 });
